vmwrapper handlers message application
- this is required since the vmservice and must be able to create a new
vm via rpc
